// HelpDeskFE/HelpDeskFETp/src/api/model/updateAgencyDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface UpdateAgencyDto {
  /**
     * @minLength 0
     * @maxLength 60
     */
  name: string;
  /**
     * @minLength 0
     * @maxLength 240
     */
  address: string;
  /**
     * @minLength 0
     * @maxLength 13
     */
  phoneNumber: string;
  /**
     * @minLength 0
     * @maxLength 240
     */
  email: string;
  /** @minimum 1 */
  idOrganization: number;
  isActive?: boolean;
}